backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
gspread
google-auth
google-api-python-client
oauth2client
